[mapbox layer] don't rerender if unnecessary
I noticed that as soon as I have a mapbox-layer enabled, many openlayers
operations like updating a style, moving a feature or zooming becomes
slower and the fan of my computer is constantly running. This is also the
case even if the operation has noting to do with the mapbox layer.
Without a mapbox-layer everything is super fast.

As it turns out, openlayers is calling render() on every layer if anything
changes anywhere. At the moment, this always triggers a complete rendering
in mapbox-gl even if nothing changed there. On my PC, this wastes about
10ms. If there are many OL-Layers this can happen very often.

We only have to render the mapbox-gl layer again, if something within it
changed.
This is also what others do:
https://github.com/uber/react-map-gl/blob/master/src/mapbox/mapbox.js

Fix this situation by keeping track of the most recent render run and
only rerender if something changed.

This helps operations which change OL-Layers a lot. Updating a style 
every 100ms did use about 40% CPU. Now it's nearly nothing.

Also the initial loading, zooming etc, becomes faster, if there
a many layers.